Make test prediction at end of setup (#33)
In general, it is possible for Predict to be broken, even if setup completes.
While it would be preferable to never push a model with a broken Predict, it's still useful to mitigate this risk by failing models with broken Predict during setup.

This commit makes a short test prediction at the end of setup. If the prediction fails, the model is broken and will fail setup.

It also fixes a bug where default prompt templates were not correctly used to format prompts.