Update downloader to support local weights cache
For some models, we ensure that weights are cached on nodes. This commit updates the downloader to be compatible with our caching strategy.

One side-effect of this implementation is that weights will not be re-used when cog-vllm is run locally, because weights are now downloaded to `/weights/vllm` instead of `/src/...`.
However, users can still re-use local weights by downloading them and then pointing to the local location.